Do not show the Winners for 2playsamonth initiative now
